refs #5

+++ src/main/java/com/vitor/amorim/sscon/geospatial/service/SalaryService.java
@@ -2,8 +2,10 @@ package com.vitor.amorim.sscon.geospatial.service;
 
 import com.vitor.amorim.sscon.geospatial.domain.outputs.SalaryOutput;
 
+import java.math.BigDecimal;
+
 public interface SalaryService {
     public String getSalary(Long pessoaId, SalaryOutput output);
-    public Long calculateSalaryIncrease(int yearsInCompany);
-    public Long calculateMininumSalaryQuantity(int yearsInCompany);
+    public BigDecimal calculateSalaryIncrease(int yearsInCompany);
+    public BigDecimal calculateMininumSalaryQuantity(int yearsInCompany);
 }
